Add Listener type to easily receive new updates
Previously you could call `long_poll` on the Api (former Bot) directly
to receive new updates in an easy way. This has some disadvantages,
because the Api is then used to listen and to send messages. It's save
to assume that a user of this library would call `long_poll` after
creating the Api and this function would run until the program exits.
This means that the created Api is borrowed mutably for the whole
time, which of course is super bad if you somehow want to use it in a
concurrent environment.

So the Api type is now split: A `Listener` that exists just for
listening. It doesn't share any internal state with the Api it's
coming from, so the Api can be used while the listener is borrowed.